fixed bug in error checking function as described on gauchospace

+++ hw3/innerproduct.cpp
@@ -24,7 +24,7 @@ double rec_cilkified(double * a, double * b, int n)
 
   if(n <= COARSENESS) // go serial
   {
-    suma = std::inner_product(a, a+n, b, 0);
+    suma = std::inner_product(a, a+n, b, (double)0);
     return suma;
   }
   else  // split and recurse
@@ -103,12 +103,12 @@ int inn_prod_driver(int n)
     	std::random_shuffle(a, a + n);
 	std::random_shuffle(b, b + n);
 
-	double seqresult = std::inner_product(a, a+n, b, 0);	
+	double seqresult = std::inner_product(a, a+n, b, (double)0);	
 
 	long t1 = example_get_time();
 	for(int i=0; i< ITERS; ++i)
 	{
-		seqresult = std::inner_product(a, a+n, b, 0);	
+		seqresult = std::inner_product(a, a+n, b, (double)0);	
 	}
 	long t2 = example_get_time();
